+++ .gitignore
@@ -45,6 +45,7 @@ techUI*
 # Symbolic link
 examples/with-antd4/src/App.tsx
 examples/with-antd4/src/style.css
+examples/with-antd4/src/WithI18n.tsx
 
 # turbo
 .turbo
